VBA Modules as README.md
I created 4 Modules to break up the script to keep in organized then condensed it & added it to the readme file as a whole. I took screen shots of every module and added each to folders 1-4 in the screen shots folder. I then added the .bas files to Modules folder. The finished Excel is in the Excel File Folder.